// src/apis/models/FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroups_Dtos_GetPacsDeviceAccessGroupForEditOutput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroupItemss_Dtos_PacsPacsDeviceAccessGroupItemsEditDto } from './FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroupItemss_Dtos_PacsPacsDeviceAccessGroupItemsEditDto';
import type { FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroups_Dtos_PacsDeviceAccessGroupEditDto } from './FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroups_Dtos_PacsDeviceAccessGroupEditDto';
export type FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroups_Dtos_GetPacsDeviceAccessGroupForEditOutput = {
    deviceAccessGroup: FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroups_Dtos_PacsDeviceAccessGroupEditDto;
    items?: Array<FireBytes_Unified_Wave_Pacs_PacsDeviceAccessGroupItemss_Dtos_PacsPacsDeviceAccessGroupItemsEditDto> | null;
};